EmptyProcessConfigCreator moved to utils, tests fixes: added application.conf in model-util
