+++ engine/flink/model-util/src/test/resources/application.conf
@@ -1,0 +1,11 @@
+checkpointConfig {
+    checkpointInterval: 10s
+}
+timeout: 10s
+rocksDB {
+  checkpointDataUri: "file:///tmp/flinkRocksDB"
+}
+asyncExecutionConfig {
+  bufferSize: 200
+  workers: 8
+}
