+++ engine/flink/process/src/test/scala/pl/touk/nussknacker/engine/util/service/query/ServiceQueryOpenCloseSpec.scala
@@ -6,7 +6,8 @@ import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.process.{ProcessObjectDependencies, WithCategories}
 import pl.touk.nussknacker.engine.api.test.InvocationCollectors.ServiceInvocationCollector
 import pl.touk.nussknacker.engine.flink.util.service.TimeMeasuringService
-import pl.touk.nussknacker.engine.testing.{EmptyProcessConfigCreator, LocalModelData}
+import pl.touk.nussknacker.engine.testing.LocalModelData
+import pl.touk.nussknacker.engine.util.process.EmptyProcessConfigCreator
 import pl.touk.nussknacker.engine.util.service.GenericTimeMeasuringService
 import pl.touk.nussknacker.test.PatientScalaFutures
