+++ engine/interpreter/src/test/scala/pl/touk/nussknacker/engine/ModelDataTest.scala
@@ -4,7 +4,8 @@ import java.util.Collections
 
 import com.typesafe.config.ConfigFactory
 import org.scalatest.{FunSuite, Matchers}
-import pl.touk.nussknacker.engine.testing.{EmptyProcessConfigCreator, LocalModelData}
+import pl.touk.nussknacker.engine.testing.LocalModelData
+import pl.touk.nussknacker.engine.util.process.EmptyProcessConfigCreator
 
 class ModelDataTest extends FunSuite with Matchers {
