+++ engine/interpreter/src/test/scala/pl/touk/nussknacker/engine/util/service/query/ExpressionServiceQuerySpec.scala
@@ -4,7 +4,8 @@ import com.typesafe.config.ConfigFactory
 import org.scalatest.{FlatSpec, Matchers}
 import pl.touk.nussknacker.engine.api.Service
 import pl.touk.nussknacker.engine.graph.expression.Expression
-import pl.touk.nussknacker.engine.testing.{EmptyProcessConfigCreator, LocalModelData}
+import pl.touk.nussknacker.engine.testing.LocalModelData
+import pl.touk.nussknacker.engine.util.process.EmptyProcessConfigCreator
 import pl.touk.nussknacker.engine.util.service.query.QueryServiceTesting.CreateQuery
 import pl.touk.nussknacker.engine.util.service.query.ServiceQuerySpec.ConcatService
 import pl.touk.nussknacker.test.PatientScalaFutures
