+++ engine/interpreter/src/test/scala/pl/touk/nussknacker/engine/util/service/query/ServiceQuerySpec.scala
@@ -5,7 +5,8 @@ import com.typesafe.config.ConfigFactory
 import org.scalatest.{FunSuite, Matchers}
 import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.process.{ProcessObjectDependencies, WithCategories}
-import pl.touk.nussknacker.engine.testing.{EmptyProcessConfigCreator, LocalModelData}
+import pl.touk.nussknacker.engine.testing.LocalModelData
+import pl.touk.nussknacker.engine.util.process.EmptyProcessConfigCreator
 import pl.touk.nussknacker.engine.util.service.query.ServiceQuery.QueryResult
 import pl.touk.nussknacker.test.PatientScalaFutures
