+++ engine/util/src/main/scala/pl/touk/nussknacker/engine/util/process/EmptyProcessConfigCreator.scala
@@ -1,4 +1,4 @@
-package pl.touk.nussknacker.engine.testing
+package pl.touk.nussknacker.engine.util.process
 
 import pl.touk.nussknacker.engine.api.exception.{EspExceptionHandler, ExceptionHandlerFactory}
 import pl.touk.nussknacker.engine.api.process.{ProcessObjectDependencies, _}
